+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# pkgcache development version
+
 # pkgcache 1.1.1
 
 * `package_cache` now does not fail if the web server does not send an