+++ NEWS.md
@@ -3,6 +3,9 @@
 
 * Failed metadata downloads now do not trigger metadata updates (#52).
 
+* New `bioc_release_version()`, `bioc_devel_version()`, `bioc_repos()`
+  helper functions to deal with Bioconductor repositories.
+
 # pkgcache 1.1.1
 
 * `package_cache` now does not fail if the web server does not send an