+++ NEWS.md
@@ -1,6 +1,12 @@
 
 # pkgcache development version
 
+* New `repo_add()`, `repo_get()`, `repo_resolve()` and `with_repo()`
+  functions to query and manipulate repositories.
+
+* `meta_cache_*()` functions now handle `getOption("repos")` changes
+  correctly.
+
 * Failed metadata downloads now do not trigger metadata updates (#52).
 
 * New `bioc_release_version()`, `bioc_devel_version()`, `bioc_repos()`