+++ NEWS.md
@@ -12,6 +12,9 @@
 * New `bioc_release_version()`, `bioc_devel_version()`, `bioc_repos()`
   helper functions to deal with Bioconductor repositories.
 
+* Metadata cache functions, e.g. `meta_cache_deps()` etc. now allow
+  specifying the dependency types in all lowercase (#54).
+
 # pkgcache 1.1.1
 
 * `package_cache` now does not fail if the web server does not send an